requirements.txt
kubernetes==11.0.0
jupyterhub==1.4.1
jupyterhub-kubespawner==1.0.0
tornado==5.1.1
git+https://github.com/jupyterhub/oauthenticator.git@93284e024c298bdff80d6e573361c33c1ed9cd21
git+https://github.com/jupyterhub/wrapspawner.git@5f2b7075f77d0c1c49066682a8e8adad0dab76db
jupyterhub-tmpauthenticator==0.6
jupyterhub-ldapauthenticator==1.2.2
psycopg2==2.8.6
openshift==0.11.2
wrapt==1.12.1
python-ldap